+++ Mod/CHANGELOG.md
@@ -1,6 +1,13 @@
 
 #Changelog
 
+## 0.2.1
+- Updated typo in description
 
-# 0.1.0
+
+## 0.2.0
+- Changed default Shattered Abodes stage to stage 2
+- Added support for adding Reformed Altar and Treeborn Colony to the regular path
+
+## 0.1.0
 Initial Release